\name{Rdtq-package}
\alias{Rdtq-package}
\alias{Rdtq}
\docType{package}
\title{
\packageTitle{Rdtq}
}
\description{
\packageDescription{Rdtq}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rdtq}
\packageIndices{Rdtq}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Rdtq}

Maintainer: \packageMaintainer{Rdtq}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
